libsepol,secilc,policycoreutils: add unprivileged sandboxing capability
This patch provides support in the selinux userspace for the
unprivileged sandboxing capability introduced by the previous kernel
patch.  As with that patch, this is a RFC on an early prototype of a
new capability to support unprivileged sandboxing.  See the kernel
patch description for a fuller description of the background, goals,
and approach.

This patch adds libsepol support for sandbox policies including deny
statements and it adds a secilc option for compiling a cil policy into
a binary sandbox file. It also adds a new utility called sebox that
allows a user to load and unload sandbox policies to the kernel.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>