libselinux: add selinux_unshare()
Provide a selinux_unshare() wrapper for /sys/fs/selinux/unshare and
other required processing when unsharing the SELinux namespace.

Add a selinux_unshare() interface to unshare the SELinux namespace,
and a unshareselinux utility to run a shell or command in its own
SELinux and mount namespaces. The selinux_unshare() interface expects
the caller to have already unshared its mount namespace and created a
MS_REC|MS_PRIVATE mount of / prior to invoking it so that it can
freely modify the selinuxfs mount as needed by the unshare
operation. The unshareselinux utility demonstrates how to do this
prior to calling selinux_unshare(). Upon a successful return from
selinux_unshare(), the SELinux namespace will be unshared and there
will be no selinuxfs mount on /sys/fs/selinux.  The caller can then
proceed to mount a new selinuxfs filesystem private to the new
namespace, load a policy, set enforcing mode, etc, as is commonly
handled by init/systemd during boot.

If/when this gets merged, the new selinux_unshare() symbol should be
moved to its own version in libselinux.map.

Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>