checkpolicy/tests: Modify tests to check handling of initial sids
For policy_allonce.conf and policy_allonce_xen.conf declare the
first three initial sids, but only assign a context to the first
and third. This will cause the second initial sid to be dropped
from the binary policy and cause the handling of a gap in the
initial sids to be tested.

Update the expected and expected_opt policies to reflect the new
expected resulting policies.

Signed-off-by: James Carter <jwcart2@gmail.com>